Add a button for each top-level content folder
These buttons, whose labels are the 'topics' that users will filter by,
should have their state recorded after users add or remove them. This
recording will happen in the global scope.

Buttons will be styled as in the compass-extras.css file.

The buttons are not yet attached to Javascript callbacks.